change cuePoint to long array, and adding two dependency injection for adinterface, and also fixing the mid-roll exit out of application twice causing movie to resume from begining

+++ lib/src/main/java/com/tubitv/media/fsm/callback/CuePointCallBack.java
@@ -6,7 +6,7 @@ package com.tubitv.media.fsm.callback;
  */
 public interface CuePointCallBack {
 
-    void onCuePointReceived(int[] quePoints);
+    void onCuePointReceived(long[] quePoints);
 
     void onCuePointError();
 }
